pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "nadeshiko-dev-tools"
version = "0.1.0"
description = "CLI utilities for anime media processing and segment generation"
requires-python = ">=3.11,<3.14"
dependencies = [
    "babelfish>=0.6.1",
    "boto3>=1.35.0",
    "deepl>=1.27.0",
    "ffmpeg-python>=0.2",
    "ffsubsync>=0.4.19",
    "guessit>=3.8.0",
    "jaconvv2>=0.4",
    "langdetect>=1.0.9",
    "pysubs2>=1.8.0",
    "python-anilist>=1.0.9",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "questionary>=2.1.1",
    "requests>=2.32.5",
    "rich>=14.3.1",
    "nadeshiko-internal-sdk>=1.5.0.dev1772172225",
    "praat-parselmouth>=0.4.4",
]

[project.optional-dependencies]
nsfw = [
    "onnxruntime>=1.17.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "Pillow>=10.0.0",
    "huggingface-hub>=0.20.0",
]
nsfw-gpu = [
    "onnxruntime-gpu>=1.17.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "Pillow>=10.0.0",
    "huggingface-hub>=0.20.0",
]
pitch = [
    "praat-parselmouth>=0.4.4",
    "demucs>=4.0.0",
    "torch>=2.0.0",
]
mora = [
    "sudachipy>=0.6.8",
    "sudachidict-full>=20240109",
    "pyopenjtalk>=0.3.0",
]
tokenizer = [
    "sudachipy>=0.6.8",
    "sudachidict-full>=20240109",
    "fugashi>=1.3.0",
    "unidic>=1.1.0",
]

[project.scripts]
media-sub-splitter = "nadeshiko_dev_tools.media_sub_splitter.__main__:main"
assets-uploader = "nadeshiko_dev_tools.assets_uploader.__main__:main"
format-migrator = "nadeshiko_dev_tools.migrator.__main__:main"
post-upload = "nadeshiko_dev_tools.post_upload:main"
nsfw-tagger = "nadeshiko_dev_tools.nsfw_tagger.__main__:main"
pitch-extractor = "nadeshiko_dev_tools.pitch_extractor.__main__:main"
mora-aligner = "nadeshiko_dev_tools.mora_aligner.__main__:main"
tokenizer = "nadeshiko_dev_tools.tokenizer.__main__:main"

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "pytest>=9.0.2",
    "pytest-snapshot>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
# Remove unused imports and variables automatically
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
